pyproject.toml
[project]
name = "prompt-caching-basics"
version = "1.0.0"
description = "Demonstration of prompt caching to reduce AI API costs by 90%"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]